agent/requirements_similarity.txt
praw>=7.7.0
google-genai>=0.8.0
scikit-learn>=1.3.0
numpy>=1.24.0